[BUGFIX] Change TCA of mail plugin field to fetch less records
Change tx_powermail_domain_model_mail.plugin TCA configuration so
that only the selected related tt_content record is fetched.

The previous TCA configuration for tx_powermail_domain_model_mail.plugin
resulted in DB queries with large result sets on large sites since
all tt_content records were returned (including bodytext field).
A lot of this was not necessary as the non-plugin records were
fetched as well. Additionally, the field is readOnly, so the results
were not displayed.

If the field is to be changed to non-readOnly, the query could be
restricted to fetch only plugin fields.

Resolves: #15